+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Changed
+- Updated Luau dependency from version 0.688 to [0.702](https://github.com/luau-lang/luau/releases/tag/0.702)
+
 ## [0.5.0] - 2025-08-22
 
 ### Added