+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.6.0] - 2025-12-03
+
 ### Changed
 - Updated Luau dependency from version 0.688 to [0.702](https://github.com/luau-lang/luau/releases/tag/0.702)
 - Raw Luau C API is now re-exported from the library via the `c` module